chore: refactor NativeScreenshot class
