+++ app/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'com.android.application'
-apply plugin: 'spoon'
 apply plugin: 'jacoco'
 jacoco {
     toolVersion='0.7.5.201505241946'
@@ -7,10 +6,6 @@ jacoco {
 buildscript {
     repositories {
         mavenCentral()
-    }
-
-    dependencies {
-        classpath 'com.stanfy.spoon:spoon-gradle-plugin:1.2.2'
     }
 }
 
@@ -46,13 +41,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
-    androidTestImplementation 'com.squareup.spoon:spoon-client:1.7.1'
     androidTestImplementation 'androidx.test:rules:1.1.1'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     testImplementation 'junit:junit:4.12'
-}
-
-spoon {
-    debug = true
-    adbTimeout = 30
 }
